myproject/requirements.txt
cffi==1.14.0
click==7.1.2
cryptography==2.9.2
Flask==1.1.2
Flask-SQLAlchemy==2.4.3
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
mysqlclient==1.4.6
pycparser==2.20
PyJWT==1.7.1
six==1.15.0
SQLAlchemy==1.3.17
Werkzeug==1.0.1
